pyproject.toml
[project]
name = "perturbscope-gpt"
version = "0.1.0"
description = "Local-first AI4Bio project for single-cell perturbation response prediction and target prioritization."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
  "anndata>=0.10",
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "scanpy>=1.10",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "streamlit>=1.33",
  "torch>=2.2",
  "xgboost>=2.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.uv]
package = false

